requirements.txt
quart
openai
python-dotenv
quart-wtf
sqlalchemy
wtforms
quart-auth
stripe
quart-mail
PyJWT
email-validator
ics
langchain
langchain-community
google-search-results
bs4
tiktoken
faiss-cpu
hypercorn
pymysql
alembic
cryptography